pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vitruvian"
version = "0.1.0"
description = "Bio-inspired coding agent built on mini-swe-agent"
requires-python = ">=3.12"
dependencies = [
    "mini-swe-agent>=2.2.0",
    "rank_bm25>=0.2.2",
    "PyYAML>=6.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0"
]

[project.scripts]
vitruvian = "vitruvian.run:app"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["vitruvian*"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["vitruvian"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
